apps/api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
sqlalchemy
asyncpg
psycopg[binary]
pgvector
redis
rq
boto3
jinja2
weasyprint
python-multipart
python-docx
docxtpl
pandas
openpyxl
xlsxwriter
httpx
tenacity
structlog
openai>=1.42.0
google-auth
google-auth-oauthlib
google-api-python-client
msal
cryptography
